% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_effect_ORSR.R
\name{calculate_effect_ORSR}
\alias{calculate_effect_ORSR}
\title{Calculate differences between OR and SR for several simulation runs}
\usage{
calculate_effect_ORSR(results_list, indices, param_combinations)
}
\arguments{
\item{results_list}{list(data.frame) a list of data.frames with mean and SE
reading times}

\item{indices}{list(numeric[]) a list of indices vectors; the first list item
of the output of \code{\link{get_reps_same_params}}}

\item{param_combinations}{character[] a vector containing unique parameter
value combinations of a search-param-space-em simulation run}
}
\value{
effects \code{numeric[]} a vector of differences between mean OR and
SR reading times
}
\description{
At the moment specialised to the analysis of experiment 1 of
Grodner & Gibson (2004). Calculates the difference between the mean reading
times of object relative (OR) and subject relative (SR) sentences. Also, at
the moment, takes the average of several simulation runs of the same parameter
value combination, if the indices vector is longer than one.
}

