% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_reading_measures.R
\name{calculate_reading_measures}
\alias{calculate_reading_measures}
\title{Calculate reading measures for ACT-R simulations}
\usage{
calculate_reading_measures(fixation_list)
}
\arguments{
\item{fixation_list}{list(data.frame) with fixation output of ACT-R simulations}
}
\value{
\code{data_list} list(list(data.frame, data.frame)), containing
data.frames with reading measures for every region.
}
\description{
Use \code{\link{em2}} to calculate reading measures for the
fixation output of ACT-R simulations.
}
\seealso{
\code{\link{get_data}} for creating a \code{fixation_list} from the
working directory.
}

