% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_main_emb_V.R
\name{extract_main_emb_V}
\alias{extract_main_emb_V}
\title{Extract data for main and embedded verb regions}
\usage{
extract_main_emb_V(data_list)
}
\arguments{
\item{data_list}{list(data.frame) containing fixation output of an ACT-R
simulation.}
}
\value{
\code{data_list} \code{list(list(data.frame, data.frame))}, containing
data.frames with reading measures for the subset regions
}
\description{
Use \code{dplyr::filter} to subset the data of a
\code{data_list} to include only the lines for main verb and embedded
verb regions. Currently only works for experiment 1 of Grodner & Gibson
(2005) (\code{gg-exp1} in ACT-R).
}
\seealso{
\code{\link{get_data}} for creating a \code{data_list} from given
fixation files.
}

