% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Reading data files for ACT-R paramsearch}
\usage{
get_data(path = ".", paramsearch)
}
\arguments{
\item{path}{string of the path of the fixation files directory}

\item{paramsearch}{data.frame containing paramsearch metainformation}
}
\value{
all_fixations \code{list(list(data.frame, data.frame))} with fixations of the ACT-R paramsearch simulations.
}
\description{
Read all fixation files from an ACT-R paramsearch and create a list of
data.frames from them.
}
\details{
This is currently only fit to output of the ACT-R function
\code{run-param-search-em} and the output it generates for experiment 1 of
Grodner & Gibson (2005). A more flexible framework will be added later.
}
\seealso{
\code{\link{get_parameters}} for creating a paramsearch data.frame
}

