% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters from paramsearch.txt}
\usage{
get_parameters(path = ".", file = "paramsearch.txt")
}
\arguments{
\item{path}{string, the path to where \code{paramsearch.txt} is located; with no "/" in the end}

\item{file}{string, the name of the file (mostly \code{paramsearch.txt})}
}
\value{
paramsearch data.frame containing paramsearch metainformation.
}
\description{
Returns a dataframe containing the information in paramsearch.txt, i.e., the
running number of the simulation set, the name of the experiment, the name of
the fixations file and the parameter values of the simulation. If \code{paramsearch.txt}
is in the current working directory, this function can be called without any
arguments.
}

