% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reps_same_params.R
\name{get_reps_same_params}
\alias{get_reps_same_params}
\title{Get indices for iterations with same parameter value combinations}
\usage{
get_reps_same_params(results_list)
}
\arguments{
\item{results_list}{list(data.frame), containing means and standard
errors by region and condition, in one dataframe per simulation set}
}
\value{
\code{indices_param_combinations} list(list(integer[]), character[]) containing [[1]]: n vectors of indices grouping
iterations with identical parameter value combinations; and [[2]]: a character vector containing the unique
parameter value combinations in the \code{results_list}
}
\description{
For the analysis of multiple sets of \code{paramsearch} experiments
in ACT-R (e.g. for model checking), this function returns indices of all the
iterations that share common parameter value combinations, so that they can be
easily plotted or analysed otherwise.
}

