% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_region.R
\name{plot_by_region}
\alias{plot_by_region}
\title{Plot total reading time by region of interest and condition}
\usage{
plot_by_region(data, maintitle, label_x = "Region",
  label_y = "Total Reading Time (ms)", limit_y = c(200, 500), d = 0.2,
  legendpos = c(0, 0.2))
}
\arguments{
\item{data}{list(data.frame), a \code{results_list} as returned by \code{\link{mean_se_by_region}}}

\item{maintitle}{string, the plot main title}

\item{label_x}{string, the label of the x axis}

\item{label_y}{string, the label of the y axis}

\item{limit_y}{2x1 numeric vector, the limits of the y-axis (x-axis is not that important because it's
automatically assigned the regions, but y represents reading times)}

\item{d}{numeric, the offset for \code{position_dodge}}

\item{legendpos, }{2x1 numeric vector, the position of the legend}
}
\description{
Use \code{ggplot2} to plot the mean total reading time for a
region of interest and experimental condition, including error bars.
}

