% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Wrapper for ACTRPRedictions2016 preprocessing functions}
\usage{
preprocess_data(path = ".")
}
\arguments{
\item{path}{string, the directory where the \code{paramsearch.txt} and fixation files are located}
}
\value{
\code{results_list} list(list(data.frame, data.frame)), containing means and standard
errors by region and condition, in one dataframe per simulation set
}
\description{
This function performs all preprocessing steps that are necessary for
visualising data simulated via the ACT-R sentence processing model, for the
first experiment in Grodner & Gibson (2005). This will be generalised in the future
to fit more experiments.
}

