% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Standard error of the mean}
\usage{
standard_error(values)
}
\arguments{
\item{values}{numeric vector}
}
\value{
standard_error_mean (numeric) The standard error of the mean of the
input vector.
}
\description{
Calculate the standard error of the mean.
}

